Fix doubleclick.
When doubleclicking, it was generating multiple loading images (one from
the doubleclick event, and some from the mousedown/mouseup). As there
was only one `loadingImg` variable used, the content was overloaded. One
loaded was firing but the loadingImg was already updated to an element
with the image not loaded, resulting in no drawing.

Also:
 - Remove unused `targetURL`.
 - Discard updates from mouseup if there was no movement.