Revamp data change detection.
Before, a bug was recreating the elements all the time. That had the
consequence to update the value visible in the UI.

Now, it is does not recreate the elements. However, lit-element only
detects changes on value, not on nested data, which does not work with
containers (or any advanced structure). To refresh the content, this
instead listen on events from the container to force a repaint.